Experiment2/PGTNet4SSD/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.6
aiosignal==1.3.1
attrs==24.2.0
axial_positional_embedding==0.2.1
brotli
certifi
charset-normalizer
click==8.1.7
contourpy==1.3.1
cvxopt==1.3.2
cycler==0.12.1
deprecation==2.1.0
docker-pycreds==0.4.0
einops==0.8.0
filelock
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
gmpy2
graphviz==0.20.3
idna
intervaltree==3.1.0
Jinja2
joblib==1.4.2
kiwisolver==1.4.7
lightning==2.4.0
lightning-utilities==0.11.9
littleutils==0.2.4
local-attention==1.9.15
lxml==5.3.0
MarkupSafe
matplotlib==3.9.2
mkl_fft
mkl_random
mpmath
multidict==6.1.0
networkx
numpy
ogb==1.3.6
outdated==0.2.2
packaging==24.2
pandas==2.2.3
performer-pytorch==1.1.4
pillow
platformdirs==4.3.6
pm4py==2.7.12.1
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
pydotplus==2.0.2
pyparsing==3.2.0
PySocks
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
pytz==2024.2
PyYAML
requests
scikit-learn==1.5.2
scipy==1.14.1
sentry-sdk==2.18.0
setproctitle==1.3.4
six==1.16.0
smmap==5.0.1
sortedcontainers==2.4.0
sympy
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
torch==2.4.0
torch-geometric==2.6.1
torch_scatter
torchaudio==2.4.0
torchmetrics==1.6.0
torchvision==0.19.0
tqdm==4.67.0
typing_extensions
tzdata==2024.2
urllib3
wandb==0.18.7
yacs==0.1.8
yarl==1.17.2
